+++ mobile/android/android-components/components/lib/crash/src/test/java/mozilla/components/lib/crash/BreadcrumbTest.kt
@@ -52,12 +52,14 @@ class BreadcrumbTest {
             ),
         )
 
-        assertEquals(reporter.crashBreadcrumbs.elementAt(0).message, testMessage)
-        assertEquals(reporter.crashBreadcrumbs.elementAt(0).data, testData)
-        assertEquals(reporter.crashBreadcrumbs.elementAt(0).category, testCategory)
-        assertEquals(reporter.crashBreadcrumbs.elementAt(0).level, testLevel)
-        assertEquals(reporter.crashBreadcrumbs.elementAt(0).type, testType)
-        assertNotNull(reporter.crashBreadcrumbs.elementAt(0).date)
+        reporter.crashBreadcrumbsCopy().elementAt(0).let {
+            assertEquals(it.message, testMessage)
+            assertEquals(it.data, testData)
+            assertEquals(it.category, testCategory)
+            assertEquals(it.level, testLevel)
+            assertEquals(it.type, testType)
+            assertNotNull(it.date)
+        }
     }
 
     @Test
@@ -85,7 +87,7 @@ class BreadcrumbTest {
                 testType,
             ),
         )
-        assertEquals(reporter.crashBreadcrumbs.size, 1)
+        assertEquals(reporter.crashBreadcrumbsCopy().size, 1)
 
         reporter.recordCrashBreadcrumb(
             Breadcrumb(
@@ -96,7 +98,7 @@ class BreadcrumbTest {
                 testType,
             ),
         )
-        assertEquals(reporter.crashBreadcrumbs.size, 2)
+        assertEquals(reporter.crashBreadcrumbsCopy().size, 2)
 
         reporter.recordCrashBreadcrumb(
             Breadcrumb(
@@ -107,7 +109,7 @@ class BreadcrumbTest {
                 testType,
             ),
         )
-        assertEquals(reporter.crashBreadcrumbs.size, 3)
+        assertEquals(reporter.crashBreadcrumbsCopy().size, 3)
     }
 
     @Test
@@ -140,8 +142,10 @@ class BreadcrumbTest {
         sleep(100) /* make sure time elapsed */
         val afterDate = Date()
 
-        assertTrue(reporter.crashBreadcrumbs.elementAt(0).date.after(beginDate))
-        assertTrue(reporter.crashBreadcrumbs.elementAt(0).date.before(afterDate))
+        reporter.crashBreadcrumbsCopy().elementAt(0).let {
+            assertTrue(it.date.after(beginDate))
+            assertTrue(it.date.before(afterDate))
+        }
 
         val date = Date()
         reporter.recordCrashBreadcrumb(
@@ -154,7 +158,10 @@ class BreadcrumbTest {
                 date,
             ),
         )
-        assertEquals(reporter.crashBreadcrumbs.elementAt(1).date.compareTo(date), 0)
+
+        reporter.crashBreadcrumbsCopy().elementAt(1).let {
+            assertEquals(it.date.compareTo(date), 0)
+        }
     }
 
     @Test